drop src folder
