+++ libs/base/ui/overlays/dropdown/index.ts
@@ -1,1 +1,3 @@
-export * from './src';
+export * from './dropdown.component';
+export * from './dropdown.model';
+export * from './dropdown.styles';
