--- libs/base/ui/overlays/dropdown/src/index.ts
@@ -1,3 +1,0 @@
-export * from './dropdown.component';
-export * from './dropdown.model';
-export * from './dropdown.styles';
