+++ libs/base/ui/overlays/dropdown/stories/Static/trigger-icon-size.stories.ts
@@ -3,7 +3,7 @@ import { IconTypes } from '@spryker-oryx/ui/icon';
 import { Size } from '@spryker-oryx/utilities';
 import { Meta, Story } from '@storybook/web-components';
 import { TemplateResult, html } from 'lit';
-import { storybookPrefix } from '../../../../../.constants';
+import { storybookPrefix } from '../../../../.constants';
 
 export default {
   title: `${storybookPrefix}/Overlays/Dropdown/Static`,
