+++ libs/base/ui/overlays/dropdown/stories/Static/vertical.stories.ts
@@ -1,7 +1,7 @@
 import { OverlaysDecorator } from '@/tools/storybook';
 import { Meta, Story } from '@storybook/web-components';
-import { html, TemplateResult } from 'lit';
-import { storybookPrefix } from '../../../../../.constants';
+import { TemplateResult, html } from 'lit';
+import { storybookPrefix } from '../../../../.constants';
 import { Position } from '../../dropdown.model';
 import '../utils';
