+++ libs/base/ui/overlays/dropdown/stories/demo.stories.ts
@@ -3,7 +3,7 @@ import { Size } from '@spryker-oryx/utilities';
 import { Story } from '@storybook/web-components';
 import { TemplateResult, html } from 'lit';
 import { when } from 'lit/directives/when.js';
-import { storybookPrefix } from '../../../../.constants';
+import { storybookPrefix } from '../../../.constants';
 import { Position } from '../dropdown.model';
 import { renderCustomContent, renderOptions } from './utils';
